Move MetricChart.svelte back to tauri-app
