+++ packages/ui-components/src/lib/index.ts
@@ -47,7 +47,6 @@ export { default as IconTelegram } from './components/IconTelegram.svelte';
 export { default as IconWalletConnect } from './components/IconWalletConnect.svelte';
 export { default as IconWarning } from './components/IconWarning.svelte';
 export { default as InputToken } from './components/input/InputToken.svelte';
-export { default as MetricChart } from './components/charts/MetricChart.svelte';
 
 //Types
 export type { AppStoresInterface } from './types/appStores.ts';
