+++ tauri-app/src/lib/components/Charts.svelte
@@ -3,7 +3,7 @@
   import type { ChartData } from '$lib/typeshare/config';
   import { transformDataForPlot } from '$lib/utils/chartData';
   import { sortBy } from 'lodash';
-  import { MetricChart } from '@rainlanguage/ui-components';
+  import MetricChart from './MetricChart.svelte';
   export let chartData: ChartData;
 </script>
